+++ apps/storybook/stories/components/data-list/DataList.mdx
@@ -1,0 +1,8 @@
+import { Canvas, Controls, Meta, Story } from "@storybook/blocks";
+import * as DataListStories from "./DataList.stories";
+import { DataListDocs } from "./DataListDocs";
+
+<Meta of={DataListStories} />
+<DataListDocs />
+<Canvas of={DataListStories.Preview} />
+<Controls of={DataListStories.Preview} />
