+++ apps/storybook/stories/components/datepicker/Datepicker.mdx
@@ -1,0 +1,8 @@
+import { Canvas, Controls, Meta } from "@storybook/blocks";
+import * as DatepickerStories from "./Datepicker.stories";
+import { DatepickerDocs } from "./DatepickerDocs";
+
+<Meta of={DatepickerStories} />
+<DatepickerDocs />
+<Canvas of={DatepickerStories.Preview} />
+<Controls of={DatepickerStories.Preview} />
