+++ apps/storybook/stories/components/daterangepicker/DateRangePicker.mdx
@@ -1,12 +1,10 @@
 
 import { Canvas, Controls, Meta } from '@storybook/blocks';
-import { DateRangePicker } from './DateRangePicker';
 import * as DateRangePickerStories from './DateRangePicker.stories';
+import { DateRangePickerDocs } from './DateRangePickerDocs';
 
 <Meta of={DateRangePickerStories} />
-
-<DateRangePicker />
-<Canvas of={DateRangePickerStories.Preview}
-  />
+<DateRangePickerDocs />
+<Canvas of={DateRangePickerStories.Preview} />
 <Controls of={DateRangePickerStories.Preview} />
