+++ apps/storybook/stories/components/dialog/Dialog.mdx
@@ -1,0 +1,10 @@
+import { Canvas, Controls, Meta } from "@storybook/blocks";
+import * as DialogStories from "./Dialog.stories";
+import { DialogDocs } from "./DialogDocs";
+
+
+<Meta of={DialogStories} />
+<DialogDocs />
+<Canvas of={DialogStories.Preview} />
+<Controls of={DialogStories.Preview} />
+
