+++ apps/storybook/stories/components/dialog/Dialog.stories.tsx
@@ -1,33 +1,26 @@
 import {
+  Box,
   Button,
-  Drawer,
-  DrawerBackdrop,
-  DrawerBody,
-  DrawerCloseTrigger,
-  DrawerContent,
-  DrawerFooter,
-  DrawerHeader,
-  DrawerPositioner,
-  DrawerProps,
-  DrawerTitle,
-  DrawerTrigger,
-  Input,
-  Drawer as KvibDrawer,
-  useDisclosure,
+  DialogActionTrigger,
+  DialogBody,
+  DialogCloseTrigger,
+  DialogContent,
+  DialogFooter,
+  DialogHeader,
+  DialogProps,
+  DialogTitle,
+  DialogTrigger,
+  Dialog as KvibModal,
 } from "@kvib/react";
 import { Meta, StoryObj } from "@storybook/react";
 
-const meta: Meta<typeof KvibDrawer> = {
-  title: "Komponenter/Drawer",
-  component: KvibDrawer,
-  parameters: {
-    docs: {
-      story: { inline: true },
-      canvas: { sourceState: "shown" },
-    },
-  },
+const meta: Meta<typeof KvibModal> = {
+  title: "Komponenter/Dialog",
+  component: KvibModal,
+
   argTypes: {
     isOpen: {
+      name: "isOpen*",
       description: "If true, the modal will be open.",
       table: {
         type: { summary: "boolean" },
@@ -42,10 +35,10 @@ const meta: Meta<typeof KvibDrawer> = {
       control: "text",
     },
     allowPinchZoom: {
+      name: "isOpen*",
       description: "Handle zoom/pinch gestures on iOS devices when scroll locking is enabled.",
       table: {
         type: { summary: "boolean" },
-        defaultValue: { summary: "false" },
       },
       control: "boolean",
     },
@@ -54,7 +47,6 @@ const meta: Meta<typeof KvibDrawer> = {
         "If true, the modal will autofocus the first enabled and interactive element within the ModalContent",
       table: {
         type: { summary: "boolean" },
-        defaultValue: { summary: "true" },
       },
       control: "boolean",
     },
@@ -62,7 +54,6 @@ const meta: Meta<typeof KvibDrawer> = {
       description: "If true, scrolling will be disabled on the body when the modal opens.",
       table: {
         type: { summary: "boolean" },
-        defaultValue: { summary: "true" },
       },
       control: "boolean",
     },
@@ -70,15 +61,13 @@ const meta: Meta<typeof KvibDrawer> = {
       description: "If true, the modal will close when the Esc key is pressed",
       table: {
         type: { summary: "boolean" },
-        defaultValue: { summary: "true" },
       },
       control: "boolean",
     },
     closeOnOverlayClick: {
-      description: "f true, the modal will close when the overlay is clicked",
+      description: "If true, the modal will close when the overlay is clicked",
       table: {
         type: { summary: "boolean" },
-        defaultValue: { summary: "true" },
       },
       control: "boolean",
     },
@@ -87,25 +76,17 @@ const meta: Meta<typeof KvibDrawer> = {
       table: {
         type: { summary: "RefObject<FocusableElement>" },
       },
-      control: "text",
-    },
-    id: {
-      description: "The id of the modal",
-      table: {
-        type: { summary: "String" },
-      },
-      control: "text",
+      control: "object",
     },
     initialFocusRef: {
       description: "The ref of element to receive focus when the modal opens.",
       table: {
         type: { summary: "RefObject<FocusableElement>" },
       },
-      control: "text",
+      control: "object",
     },
-    isFullHeight: {
-      description:
-        "If true and drawer's placement is top or bottom, the drawer will occupy the viewport height (100vh)",
+    isCentered: {
+      description: "If true, the modal will be centered on screen.",
       table: {
         type: { summary: "boolean" },
       },
@@ -116,12 +97,26 @@ const meta: Meta<typeof KvibDrawer> = {
         "Enables aggressive focus capturing within iframes. - If true: keep focus in the lock, no matter where lock is active - If false: allows focus to move outside of iframe",
       table: {
         type: { summary: "boolean" },
-        defaultValue: { summary: "false" },
       },
       control: "boolean",
     },
+    motionPreset: {
+      description: "The transition that should be used for the modal",
+      table: {
+        type: { summary: "MotionPreset" },
+        defaultValue: { summary: "scale" },
+      },
+      control: "text",
+    },
     onCloseComplete: {
       description: "Fires when all exiting nodes have completed animating out",
+      table: {
+        type: { summary: "() => void" },
+      },
+      control: "text",
+    },
+    onEsc: {
+      description: "Callback fired when the escape key is pressed and focus is within modal",
       table: {
         type: { summary: "() => void" },
       },
@@ -132,14 +127,6 @@ const meta: Meta<typeof KvibDrawer> = {
       table: {
         type: { summary: "() => void" },
       },
-      control: "text",
-    },
-    placement: {
-      description: "The placement of the drawer",
-      table: {
-        type: { summary: "bottom | top | start | end" },
-      },
-      defaultValue: { summary: "end" },
       control: "text",
     },
     portalProps: {
@@ -154,7 +141,6 @@ const meta: Meta<typeof KvibDrawer> = {
         "If true, a `padding-right` will be applied to the body element that's equal to the width of the scrollbar. This can help prevent some unpleasant flickering effect and content adjustment when the modal opens",
       table: {
         type: { summary: "boolean" },
-        defaultValue: { summary: "true" },
       },
       control: "boolean",
     },
@@ -162,17 +148,25 @@ const meta: Meta<typeof KvibDrawer> = {
       description: "If true, the modal will return focus to the element that triggered it when it closes.",
       table: {
         type: { summary: "boolean" },
-        defaultValue: { summary: "true" },
       },
       control: "boolean",
     },
-    size: {
-      description: "Size of the Drawer",
+    scrollBehavior: {
+      description:
+        "Where scroll behavior should originate. - If set to inside, scroll only occurs within the ModalBody. - If set to outside, the entire ModalContent will scroll within the viewport.",
       table: {
-        type: { summary: "xs | sm | md | lg | xl | full" },
-        defaultValue: { summary: "xs" },
+        type: { summary: "ScrollBehavior" },
+        defaultValue: { summary: "outside" },
       },
-      options: ["xs", "sm", "md", "lg", "xl", "full"],
+      control: "text",
+    },
+    size: {
+      description: "The size of the Modal",
+      table: {
+        type: { summary: "xs | sm | md | lg | xl | 2xl | 3xl | 4xl | 5xl | 6xl | full" },
+        defaultValue: { summary: "md" },
+      },
+      options: ["xs", "sm", "md", "lg", "xl", "2xl", "3xl", "4xl", "5xl", "6xl", "full"],
       control: { type: "radio" },
     },
     trapFocus: {
@@ -180,7 +174,6 @@ const meta: Meta<typeof KvibDrawer> = {
         "If false, focus lock will be disabled completely. This is useful in situations where you still need to interact with other surrounding elements. 🚨Warning: We don't recommend doing this because it hurts the accessibility of the modal, based on WAI-ARIA specifications.",
       table: {
         type: { summary: "boolean" },
-        defaultValue: { summary: "true" },
       },
       control: "boolean",
     },
@@ -189,65 +182,74 @@ const meta: Meta<typeof KvibDrawer> = {
         "A11y: If true, the siblings of the modal will have `aria-hidden` set to true so that screen readers can only see the modal. This is commonly known as making the other elements **inert**",
       table: {
         type: { summary: "boolean" },
-        defaultValue: { summary: "true" },
       },
       control: "boolean",
     },
+    colorPalette: {
+      description: "The visual color appearance of the component",
+      table: {
+        type: { summary: "green | blue | gray | red" },
+      },
+      defaultValue: { summary: "green" },
+      options: ["green", "blue", "gray", "red"],
+      control: { type: "radio" },
+    },
   },
   args: {
-    allowPinchZoom: false,
-    autoFocus: true,
-    blockScrollOnMount: true,
-    closeOnEsc: true,
-    closeOnOverlayClick: true,
-    lockFocusAcrossFrames: false,
-    placement: "end",
-    preserveScrollBarGap: true,
-    returnFocusOnClose: true,
-    size: "xs",
-    trapFocus: true,
-    useInert: true,
-    onOverlayClick: undefined,
-    onClose: undefined,
+    colorPalette: "green",
     onCloseComplete: undefined,
     onEsc: undefined,
+    onOverlayClick: undefined,
+    onClose: undefined,
   },
 };
 
 export default meta;
-type Story = StoryObj<typeof KvibDrawer>;
+type Story = StoryObj<typeof KvibModal>;
 
-export const DrawerExample = ({ ...args }) => {
-  const { open, onOpen, onClose } = useDisclosure();
+const ModalExample = ({ ...args }) => {
   return (
-    <Drawer {...args} open={open} onClose={onClose}>
-      <DrawerBackdrop />
-      <Button as={DrawerTrigger} colorPalette={args.colorPalette} onClick={onOpen}>
-        Åpne
-      </Button>
-      <DrawerPositioner>
-        <DrawerContent>
-          <DrawerCloseTrigger />
-          <DrawerHeader>
-            <DrawerTitle>Lag din konto</DrawerTitle>
-          </DrawerHeader>
+    <>
+      <KvibModal {...args}>
+        <DialogTrigger asChild>
+          <Button colorPalette={args.colorPalette}>Åpne dialog</Button>
+        </DialogTrigger>
+        <DialogContent>
+          <DialogHeader>
+            <DialogTitle>Her er en modal</DialogTitle>
+          </DialogHeader>
+          <DialogBody>
+            Modaler må kun vises etter en brukerinteraksjon, og skal ikke avbryte brukeren på noe vis.
+          </DialogBody>
 
-          <DrawerBody>
-            <Input placeholder="Skriv her..." />
-          </DrawerBody>
-
-          <DrawerFooter>
-            <Button colorPalette={args.colorPalette} variant="secondary" mr={3} onClick={onClose}>
-              Avbryt
+          <DialogFooter justifyContent="space-between">
+            <Button variant="tertiary" colorPalette={args.colorPalette}>
+              Tertiær
             </Button>
-            <Button colorPalette={args.colorPalette}>Lagre</Button>
-          </DrawerFooter>
-        </DrawerContent>
-      </DrawerPositioner>
-    </Drawer>
+            <Box>
+              <Button mr={3} variant="secondary" colorPalette={args.colorPalette}>
+                Sekundær
+              </Button>
+              <DialogActionTrigger asChild>
+                <Button colorPalette={args.colorPalette}>Primær</Button>
+              </DialogActionTrigger>
+            </Box>
+          </DialogFooter>
+          <DialogCloseTrigger />
+        </DialogContent>
+      </KvibModal>
+    </>
   );
 };
 
 export const Preview: Story = {
-  render: (args: DrawerProps) => <DrawerExample {...args} />,
+  render: (args: DialogProps) => <ModalExample {...args} />,
+  parameters: {
+    docs: {
+      source: {
+        type: "code",
+        language: "tsx",
+      },
+    },
+  },
 };
