+++ apps/storybook/stories/components/drawer/Drawer.stories.tsx
@@ -1,27 +1,32 @@
 import {
-  Box,
   Button,
-  DialogActionTrigger,
-  DialogBackdrop,
-  DialogBody,
-  DialogCloseTrigger,
-  DialogContent,
-  DialogFooter,
-  DialogHeader,
-  DialogProps,
-  DialogTitle,
-  DialogTrigger,
-  Dialog as KvibModal,
+  Drawer,
+  DrawerBody,
+  DrawerCloseTrigger,
+  DrawerContent,
+  DrawerFooter,
+  DrawerHeader,
+  DrawerPositioner,
+  DrawerProps,
+  DrawerTitle,
+  DrawerTrigger,
+  Input,
+  Drawer as KvibDrawer,
+  useDisclosure,
 } from "@kvib/react";
 import { Meta, StoryObj } from "@storybook/react";
 
-const meta: Meta<typeof KvibModal> = {
-  title: "Komponenter/Dialog",
-  component: KvibModal,
-
+const meta: Meta<typeof KvibDrawer> = {
+  title: "Komponenter/Drawer",
+  component: KvibDrawer,
+  parameters: {
+    docs: {
+      story: { inline: true },
+      canvas: { sourceState: "shown" },
+    },
+  },
   argTypes: {
     isOpen: {
-      name: "isOpen*",
       description: "If true, the modal will be open.",
       table: {
         type: { summary: "boolean" },
@@ -36,10 +41,10 @@ const meta: Meta<typeof KvibModal> = {
       control: "text",
     },
     allowPinchZoom: {
-      name: "isOpen*",
       description: "Handle zoom/pinch gestures on iOS devices when scroll locking is enabled.",
       table: {
         type: { summary: "boolean" },
+        defaultValue: { summary: "false" },
       },
       control: "boolean",
     },
@@ -48,6 +53,7 @@ const meta: Meta<typeof KvibModal> = {
         "If true, the modal will autofocus the first enabled and interactive element within the ModalContent",
       table: {
         type: { summary: "boolean" },
+        defaultValue: { summary: "true" },
       },
       control: "boolean",
     },
@@ -55,6 +61,7 @@ const meta: Meta<typeof KvibModal> = {
       description: "If true, scrolling will be disabled on the body when the modal opens.",
       table: {
         type: { summary: "boolean" },
+        defaultValue: { summary: "true" },
       },
       control: "boolean",
     },
@@ -62,13 +69,15 @@ const meta: Meta<typeof KvibModal> = {
       description: "If true, the modal will close when the Esc key is pressed",
       table: {
         type: { summary: "boolean" },
+        defaultValue: { summary: "true" },
       },
       control: "boolean",
     },
     closeOnOverlayClick: {
-      description: "If true, the modal will close when the overlay is clicked",
+      description: "f true, the modal will close when the overlay is clicked",
       table: {
         type: { summary: "boolean" },
+        defaultValue: { summary: "true" },
       },
       control: "boolean",
     },
@@ -77,17 +86,25 @@ const meta: Meta<typeof KvibModal> = {
       table: {
         type: { summary: "RefObject<FocusableElement>" },
       },
-      control: "object",
+      control: "text",
+    },
+    id: {
+      description: "The id of the modal",
+      table: {
+        type: { summary: "String" },
+      },
+      control: "text",
     },
     initialFocusRef: {
       description: "The ref of element to receive focus when the modal opens.",
       table: {
         type: { summary: "RefObject<FocusableElement>" },
       },
-      control: "object",
+      control: "text",
     },
-    isCentered: {
-      description: "If true, the modal will be centered on screen.",
+    isFullHeight: {
+      description:
+        "If true and drawer's placement is top or bottom, the drawer will occupy the viewport height (100vh)",
       table: {
         type: { summary: "boolean" },
       },
@@ -98,26 +115,12 @@ const meta: Meta<typeof KvibModal> = {
         "Enables aggressive focus capturing within iframes. - If true: keep focus in the lock, no matter where lock is active - If false: allows focus to move outside of iframe",
       table: {
         type: { summary: "boolean" },
+        defaultValue: { summary: "false" },
       },
       control: "boolean",
     },
-    motionPreset: {
-      description: "The transition that should be used for the modal",
-      table: {
-        type: { summary: "MotionPreset" },
-        defaultValue: { summary: "scale" },
-      },
-      control: "text",
-    },
     onCloseComplete: {
       description: "Fires when all exiting nodes have completed animating out",
-      table: {
-        type: { summary: "() => void" },
-      },
-      control: "text",
-    },
-    onEsc: {
-      description: "Callback fired when the escape key is pressed and focus is within modal",
       table: {
         type: { summary: "() => void" },
       },
@@ -128,6 +131,14 @@ const meta: Meta<typeof KvibModal> = {
       table: {
         type: { summary: "() => void" },
       },
+      control: "text",
+    },
+    placement: {
+      description: "The placement of the drawer",
+      table: {
+        type: { summary: "bottom | top | start | end" },
+      },
+      defaultValue: { summary: "end" },
       control: "text",
     },
     portalProps: {
@@ -142,6 +153,7 @@ const meta: Meta<typeof KvibModal> = {
         "If true, a `padding-right` will be applied to the body element that's equal to the width of the scrollbar. This can help prevent some unpleasant flickering effect and content adjustment when the modal opens",
       table: {
         type: { summary: "boolean" },
+        defaultValue: { summary: "true" },
       },
       control: "boolean",
     },
@@ -149,25 +161,17 @@ const meta: Meta<typeof KvibModal> = {
       description: "If true, the modal will return focus to the element that triggered it when it closes.",
       table: {
         type: { summary: "boolean" },
+        defaultValue: { summary: "true" },
       },
       control: "boolean",
     },
-    scrollBehavior: {
-      description:
-        "Where scroll behavior should originate. - If set to inside, scroll only occurs within the ModalBody. - If set to outside, the entire ModalContent will scroll within the viewport.",
-      table: {
-        type: { summary: "ScrollBehavior" },
-        defaultValue: { summary: "outside" },
-      },
-      control: "text",
-    },
     size: {
-      description: "The size of the Modal",
+      description: "Size of the Drawer",
       table: {
-        type: { summary: "xs | sm | md | lg | xl | 2xl | 3xl | 4xl | 5xl | 6xl | full" },
-        defaultValue: { summary: "md" },
+        type: { summary: "xs | sm | md | lg | xl | full" },
+        defaultValue: { summary: "xs" },
       },
-      options: ["xs", "sm", "md", "lg", "xl", "2xl", "3xl", "4xl", "5xl", "6xl", "full"],
+      options: ["xs", "sm", "md", "lg", "xl", "full"],
       control: { type: "radio" },
     },
     trapFocus: {
@@ -175,6 +179,7 @@ const meta: Meta<typeof KvibModal> = {
         "If false, focus lock will be disabled completely. This is useful in situations where you still need to interact with other surrounding elements. 🚨Warning: We don't recommend doing this because it hurts the accessibility of the modal, based on WAI-ARIA specifications.",
       table: {
         type: { summary: "boolean" },
+        defaultValue: { summary: "true" },
       },
       control: "boolean",
     },
@@ -183,76 +188,64 @@ const meta: Meta<typeof KvibModal> = {
         "A11y: If true, the siblings of the modal will have `aria-hidden` set to true so that screen readers can only see the modal. This is commonly known as making the other elements **inert**",
       table: {
         type: { summary: "boolean" },
+        defaultValue: { summary: "true" },
       },
       control: "boolean",
     },
-    colorPalette: {
-      description: "The visual color appearance of the component",
-      table: {
-        type: { summary: "green | blue | gray | red" },
-      },
-      defaultValue: { summary: "green" },
-      options: ["green", "blue", "gray", "red"],
-      control: { type: "radio" },
-    },
   },
   args: {
-    colorPalette: "green",
-    onCloseComplete: undefined,
-    onEsc: undefined,
+    allowPinchZoom: false,
+    autoFocus: true,
+    blockScrollOnMount: true,
+    closeOnEsc: true,
+    closeOnOverlayClick: true,
+    lockFocusAcrossFrames: false,
+    placement: "end",
+    preserveScrollBarGap: true,
+    returnFocusOnClose: true,
+    size: "xs",
+    trapFocus: true,
+    useInert: true,
     onOverlayClick: undefined,
     onClose: undefined,
+    onCloseComplete: undefined,
+    onEsc: undefined,
   },
 };
 
 export default meta;
-type Story = StoryObj<typeof KvibModal>;
+type Story = StoryObj<typeof KvibDrawer>;
 
-const ModalExample = ({ ...args }) => {
+const DrawerExample = ({ ...args }) => {
+  const { open, onOpen, onClose } = useDisclosure();
   return (
-    <>
-      <KvibModal {...args}>
-        <DialogTrigger asChild>
-          <Button colorPalette={args.colorPalette}>Åpne dialog</Button>
-        </DialogTrigger>
-
-        <DialogBackdrop />
-        <DialogContent>
-          <DialogHeader>
-            <DialogTitle>Her er en modal</DialogTitle>
-          </DialogHeader>
-          <DialogBody>
-            Modaler må kun vises etter en brukerinteraksjon, og skal ikke avbryte brukeren på noe vis.
-          </DialogBody>
-
-          <DialogFooter justifyContent="space-between">
-            <Button variant="tertiary" colorPalette={args.colorPalette}>
-              Tertiær
+    <Drawer {...args} open={open} onClose={onClose}>
+      <DrawerTrigger asChild>
+        <Button colorPalette={args.colorPalette} onClick={onOpen}>
+          Åpne
+        </Button>
+      </DrawerTrigger>
+      <DrawerPositioner>
+        <DrawerContent>
+          <DrawerCloseTrigger />
+          <DrawerHeader>
+            <DrawerTitle>Lag din konto</DrawerTitle>
+          </DrawerHeader>
+          <DrawerBody>
+            <Input placeholder="Skriv her..." />
+          </DrawerBody>
+          <DrawerFooter>
+            <Button colorPalette={args.colorPalette} variant="secondary" mr={3} onClick={onClose}>
+              Avbryt
             </Button>
-            <Box>
-              <Button mr={3} variant="secondary" colorPalette={args.colorPalette}>
-                Sekundær
-              </Button>
-              <DialogActionTrigger asChild>
-                <Button colorPalette={args.colorPalette}>Primær</Button>
-              </DialogActionTrigger>
-            </Box>
-          </DialogFooter>
-          <DialogCloseTrigger />
-        </DialogContent>
-      </KvibModal>
-    </>
+            <Button colorPalette={args.colorPalette}>Lagre</Button>
+          </DrawerFooter>
+        </DrawerContent>
+      </DrawerPositioner>
+    </Drawer>
   );
 };
 
 export const Preview: Story = {
-  render: (args: DialogProps) => <ModalExample {...args} />,
-  parameters: {
-    docs: {
-      source: {
-        type: "code",
-        language: "tsx",
-      },
-    },
-  },
+  render: (args: DrawerProps) => <DrawerExample {...args} />,
 };
