--- apps/storybook/stories/components/overlay/drawer/Drawer.mdx
@@ -1,29 +1,0 @@
-import { Card, CardBody } from "@kvib/react";
-import { Canvas, Controls, Meta, Source, Story } from "@storybook/blocks";
-import { DocsAnatomy, DocsContainer, DocsHeading, DocsStory, Feedback } from "../../../templates";
-import * as DrawerStories from "./Drawer.stories";
-
-<Meta of={DrawerStories} />
-
-<DocsContainer>
-
-# Drawer
-
-Drawer er et panel som sklir ut fra kanten av skjermen. Den kan være nyttig hvis du vil at brukeren skal se detaljer eller gjøre en oppgave
-uten å bytte side. Drawer kan komme ut fra alle kanter av skjermen.
-
-```jsx
-import {
-  Drawer,
-  DrawerBody,
-  DrawerFooter,
-  DrawerHeader,
-  DrawerOverlay,
-  DrawerContent,
-  DrawerCloseButton,
-} from "@kvib/react";
-```
-
-<Canvas of={DrawerStories.Preview} />
-<Controls of={DrawerStories.Preview} />
-</DocsContainer>
