--- apps/storybook/stories/components/overlay/modal/Dialog.mdx
@@ -1,21 +1,0 @@
-import { Canvas, Controls, Meta, Primary, Source, Story } from "@storybook/blocks";
-import { DocsAnatomy, DocsContainer, DocsHeading, DocsStory, Feedback } from "../../../templates";
-import * as DialogStories from "./Dialog.stories";
-
-
-<Meta of={DialogStories} />
-
-<DocsContainer>
-
-# Dialog
-
-En dialog er en komponent som fokuserer brukerens oppmerksomhet på informasjonen i vinduet.
-
-<br/><br/>
-
-## Props
-<Canvas of={DialogStories.Preview}/>
-<Controls of={DialogStories.Preview}/>
-
-
-</DocsContainer>
