--- apps/storybook/stories/components/skjemaelementer/datepicker/Datepicker.mdx
@@ -1,55 +1,0 @@
-import { Box, Code, Text } from "@kvib/react";
-import { Canvas, Controls, Meta, Story } from "@storybook/blocks";
-import { DocsAnatomy, DocsContainer, DocsHeading, DocsStory, Feedback } from "../../../templates";
-import * as DatepickerStories from "./Datepicker.stories";
-
-<Meta of={DatepickerStories} />
-
-# Datepicker
-
-Datepicker kombinerer et inputfelt med en popoverkalender hvor brukere kan velge dato.
-
-```jsx
-import { Datepicker } from "@kvib/react";
-```
-
-<DocsContainer>
-
-**Egnet til:**
-
-- Velge en dato
-
-**Uegnet til:**
-
-- Valg av dato langt tilbake eller frem i tid
-
-
-</DocsContainer>
-
-<DocsHeading light>Retningslinjer</DocsHeading>
-
-<DocsContainer>
-### Vurder om Datepicker er den rette løsningen for deg
-
-Datepicker gjør det enkelt for brukere å navigere gjennom datoer innenfor en bestemt tidsperiode. Imidlertid kan det oppstå utfordringer når den skal dekke lengre tidsrom. For eksempel, når brukeren skal velge et år som ligger langt frem eller tilbake i tid, kan det bli nødvendig å bla gjennom en lang liste med 50 eller flere alternativer i en nedtrekksmeny.
-
-Datovelgere er komplekse komponenter, og i noen tilfeller kan det være mer brukervennlig å la brukeren skrive inn datoen manuelt i en vanlig input. [Forskning utført av Gov.uk](https://design-system.service.gov.uk/patterns/dates/) viser at det kan være en fordel å tilby et enkelt tekstfelt fremfor en interaktiv datovelger.
-
-### Label
-
-Datepicker skal alltid ha en label. Dersom Datepicker av en eller annen grunn ikke har label må en `aria-label` legges til for å tilfredsstille tilgjengelighetskrav.
-
-### Native datepicker på mobil
-
-Datepicker har som standard satt `useNative` til true. Den proppen bestemmer om Datepicker skal bruke native dropdown-kalender på mobil ({`<`}480 px). Native Datepicker for mobil er optimalisert for touch, samtidig som brukeren allerede er vant til den.
-Med native datepicker vil ikke de følgende proppene bli brukt på mobil: `showDropdownMonthYear`, `showOutsideDays`, `showWeekNumber`, `disabledDays`. Det er anbefalt å bruke native Datepicker på mobil, men det er fullt mulig å sette `useNative={false}` for å benytte seg av alle props.
-
-### Date
-
-Datepicker returnerer datoer i [Date](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date)-format. Du må selv ta deg av konvertering til ønsket format og da evt tidssoner. I selve komponenten bruker vi [date-fns](https://date-fns.org/) for parsing.
-
-</DocsContainer>
-
-<DocsHeading light>Props</DocsHeading>
-<Canvas of={DatepickerStories.Preview} />
-<Controls of={DatepickerStories.Preview} />
