+++ apps/storybook/stories/documentation/komponentoversikt/StoryRendering.tsx
@@ -22,6 +22,9 @@ import * as ListStories from "../../components/datavisning/list/List.stories";
 import * as StatStories from "../../components/datavisning/stat/Stat.stories";
 import * as TableStories from "../../components/datavisning/table/Table.stories";
 import * as TagStories from "../../components/datavisning/tag/Tag.stories";
+import * as DatepickerStories from "../../components/datepicker/Datepicker.stories";
+import * as DialogStoriea from "../../components/dialog/Dialog.stories";
+import * as DrawerStories from "../../components/drawer/Drawer.stories";
 import * as AspectRatioStories from "../../components/Layout/AspectRatio/AspectRatio.stories";
 import * as BoxStories from "../../components/Layout/Box/Box.stories";
 import * as CenterStories from "../../components/Layout/Center/Center.stories";
@@ -37,8 +40,6 @@ import * as LinkStories from "../../components/navigasjon/link/Link.stories";
 import * as MenuStories from "../../components/navigasjon/menu/Menu.stories";
 import * as SkivNavStories from "../../components/navigasjon/skip-nav/SkipNav.stories";
 import * as TabsStories from "../../components/navigasjon/tabs/Tabs.stories";
-import * as DrawerStories from "../../components/overlay/drawer/Drawer.stories";
-import * as DialogStoriea from "../../components/overlay/modal/Dialog.stories";
 import * as PopoverStories from "../../components/overlay/popover/Popover.stories";
 import * as TooltipStories from "../../components/overlay/tooltip/Tooltip.stories";
 import * as SearchAsyncStories from "../../components/search/search-async/SearchAsync.stories";
@@ -47,7 +48,6 @@ import * as FooterStories from "../../components/sideelementer/footer/footer/Foo
 import * as HeaderStories from "../../components/sideelementer/header/Header.stories";
 import * as LogoStories from "../../components/sideelementer/logo/Logo.stories";
 import * as SeparatorStories from "../../components/sideelementer/separator/Separator.stories";
-import * as DatepickerStories from "../../components/skjemaelementer/datepicker/Datepicker.stories";
 import * as EditableStories from "../../components/skjemaelementer/editable/Editable.stories";
 import * as FormControlStories from "../../components/skjemaelementer/field/Field.stories";
 import * as FileUploadStories from "../../components/skjemaelementer/file-upload/FileUpload.stories";
