+++ packages/react/src/index.ts
@@ -13,6 +13,7 @@ export * from "./close-button";
 export * from "./collapsible";
 export * from "./color-picker";
 export * from "./color-swatch";
+export * from "./data-list";
 export * from "./datepicker";
 export * from "./dialog";
 export * from "./drawer";
